refactor: restructure Meteora service into modular architecture
- Split 900+ line meteora.service.ts into focused modules:
  - types.ts: All TypeScript interfaces and types
  - constants.ts: DLMM configuration and API endpoints
  - meteora.pools.service.ts: Pool discovery and search functions
  - meteora.position.service.ts: Position management operations
  - utils.ts: Calculation utilities and helpers
  - index.ts: Clean public API exports

- Fix Redux selector performance issue with memoized selectFilteredGroups
- Update all component imports to use new modular structure
- Maintain backward compatibility through index.ts exports
- Improve code organization and maintainability
